requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
pandas==2.2.2
playwright>=1.40.0
